# Optimize order by ts desc with filter 