--- a/core/src/main/java/io/questdb/griffin/SqlCodeGenerator.java
+++ b/core/src/main/java/io/questdb/griffin/SqlCodeGenerator.java
@@ -2783,20 +2783,13 @@ public class SqlCodeGenerator implements Mutable, Closeable {
                 }
 
                 // below code block generates index-based filter
-                //tODO: check if intrinsic model.keycolumn is not a better field to check than model
-                boolean isOrderByTimestampDesc = isOrderDescendingByDesignatedTimestampOnly(model);
-
                 final boolean intervalHitsOnlyOnePartition;
                 if (intrinsicModel.hasIntervalFilters()) {
                     RuntimeIntrinsicIntervalModel intervalModel = intrinsicModel.buildIntervalModel();
                     dfcFactory = new IntervalFwdDataFrameCursorFactory(engine, tableName, model.getTableId(), model.getTableVersion(), intervalModel, readerTimestampIndex);
                     intervalHitsOnlyOnePartition = intervalModel.allIntervalsHitOnePartition(reader.getPartitionedBy());
                 } else {
-                    if (isOrderByTimestampDesc) {
-                        dfcFactory = new FullBwdDataFrameCursorFactory(engine, tableName, model.getTableId(), model.getTableVersion()); //here
-                    } else {
-                        dfcFactory = new FullFwdDataFrameCursorFactory(engine, tableName, model.getTableId(), model.getTableVersion()); //here
-                    }
+                    dfcFactory = new FullFwdDataFrameCursorFactory(engine, tableName, model.getTableId(), model.getTableVersion());
                     intervalHitsOnlyOnePartition = false;
                 }
 
@@ -3007,8 +3000,15 @@ public class SqlCodeGenerator implements Mutable, Closeable {
                     }
                 }
 
-                RowCursorFactory rowFactory = isOrderByTimestampDesc ? new BwdDataFrameRowCursorFactory() :
-                        new DataFrameRowCursorFactory();
+                boolean isOrderByTimestampDesc = isOrderDescendingByDesignatedTimestampOnly(model);
+                RowCursorFactory rowFactory;
+
+                if (isOrderByTimestampDesc) {
+                    dfcFactory = new FullBwdDataFrameCursorFactory(engine, tableName, model.getTableId(), model.getTableVersion());
+                    rowFactory = new BwdDataFrameRowCursorFactory();
+                } else {
+                    rowFactory = new DataFrameRowCursorFactory();
+                }
 
                 model.setWhereClause(intrinsicModel.filter);
                 return new DataFrameRecordCursorFactory(