--- a/core/src/main/java/io/questdb/griffin/engine/table/CompiledFilterRecordCursor.java
+++ b/core/src/main/java/io/questdb/griffin/engine/table/CompiledFilterRecordCursor.java
@@ -27,6 +27,7 @@ package io.questdb.griffin.engine.table;
 import io.questdb.cairo.*;
 import io.questdb.cairo.sql.*;
 import io.questdb.cairo.vm.Vm;
+import io.questdb.cairo.sql.Record;
 import io.questdb.cairo.vm.api.MemoryCARW;
 import io.questdb.griffin.SqlException;
 import io.questdb.griffin.SqlExecutionContext;