--- a/core/src/main/java/io/questdb/griffin/engine/table/CompiledFilterRecordCursorFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/table/CompiledFilterRecordCursorFactory.java
@@ -62,7 +62,7 @@ public class CompiledFilterRecordCursorFactory implements RecordCursorFactory {
         this.factory = factory;
         this.filter = filter;
         this.compiledFilter = compiledFilter;
-        this.cursor = new CompiledFilterRecordCursor(configuration);
+        this.cursor = new CompiledFilterRecordCursor(configuration, hasDescendingOrder());
         this.bindVarFunctions = bindVarFunctions;
         this.bindVarMemory = Vm.getCARWInstance(configuration.getSqlJitBindVarsMemoryPageSize(),
                 configuration.getSqlJitBindVarsMemoryMaxPages(), MemoryTag.NATIVE_JIT);