--- a/core/src/main/java/io/questdb/griffin/engine/table/FilteredRecordCursorFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/table/FilteredRecordCursorFactory.java
@@ -76,4 +76,9 @@ public class FilteredRecordCursorFactory implements RecordCursorFactory {
     public boolean usesCompiledFilter() {
         return base.usesCompiledFilter();
     }
+
+    @Override
+    public boolean hasDescendingOrder() {
+        return base.hasDescendingOrder();
+    }
 }