--- a/core/src/test/java/io/questdb/cairo/OrderByDescRowSkippingTest.java
+++ b/core/src/test/java/io/questdb/cairo/OrderByDescRowSkippingTest.java
@@ -800,14 +800,12 @@ public class OrderByDescRowSkippingTest extends AbstractGriffinTest {
                         "3\t2022-01-05T07:33:20.000000Z\n" +
                         "2\t2022-01-04T03:46:40.000000Z\n" +
                         "1\t2022-01-03T00:00:00.000000Z\n",
-                "select l, ts from trips where l <=5 order by ts desc limit 5", //where l <=5    limit 5, 10
+                "select l, ts from trips where l <=5 order by ts desc limit 5",
                 null, "ts###DESC", true, false, true);
     }
 
-    @Ignore
     @Test
     public void testOrderByTimestampWithJittedFilter() throws Exception {
-        //sqlExecutionContext.setJitMode(SqlJitMode.JIT_MODE_DISABLED); use default
 
         runQueries("CREATE TABLE trips(l long, ts TIMESTAMP) timestamp(ts) partition by year;",
                 "insert into trips " +
@@ -821,8 +819,127 @@ public class OrderByDescRowSkippingTest extends AbstractGriffinTest {
                         "3\t2022-01-05T07:33:20.000000Z\n" +
                         "2\t2022-01-04T03:46:40.000000Z\n" +
                         "1\t2022-01-03T00:00:00.000000Z\n",
-                "select l, ts from trips where l <=5 order by ts desc limit 5", //where l <=5    limit 5, 10
-                null, "ts###ASC", true, false, true);
+                "select l, ts from trips where l <=5 order by ts desc limit 5",
+                null, "ts###DESC", true, false, true);
+    }
+
+    @Test
+    public void testOrderByTimestampWithComplexJittedFilter() throws Exception {
+
+        runQueries("CREATE TABLE trips(l long, ts TIMESTAMP) timestamp(ts) partition by month;",
+                "insert into trips " +
+                        "  select x," +
+                        "  timestamp_sequence(to_timestamp('2022-01-03T00:00:00', 'yyyy-MM-ddTHH:mm:ss'), 100000000000) " +
+                        "  from long_sequence(10);");
+
+        assertQuery("l\tts\n" +
+                        "5\t2022-01-07T15:06:40.000000Z\n",
+                "select l, ts from trips " +
+                        "where l <=5 and ts < to_timestamp('2022-01-08T00:00:00', 'yyyy-MM-ddTHH:mm:ss') " +
+                        "order by ts desc limit 1",
+                null, "ts###DESC", true, false, true);
+    }
+
+    @Test//triggers DeferredSingleSymbolFilterDataFrameRecordCursorFactory
+    public void testOrderByDescSelectByIndexedSymbolColumn() throws Exception {
+        runQueries("CREATE TABLE trips(l long,s symbol index capacity 10, ts TIMESTAMP) timestamp(ts) partition by month;",
+                "insert into trips " +
+                        "  select x, case when x <= 5 then 'ABC' when x <= 7 then 'DEF' else 'GHI' end," +
+                        "  timestamp_sequence(to_timestamp('2022-01-03T00:00:00', 'yyyy-MM-ddTHH:mm:ss'), 100000000000) " +
+                        "  from long_sequence(10);");
+
+        assertQuery("l\ts\tts\n" +
+                        "7\tDEF\t2022-01-09T22:40:00.000000Z\n" +
+                        "6\tDEF\t2022-01-08T18:53:20.000000Z\n",
+                "select l, s, ts from trips where s = 'DEF' order by ts desc",
+                null, null, true, false, false);
+    }
+
+    @Test
+    public void testOrderByDescWithFilterOnSubQueryRecordCursorFactory() throws Exception {
+        runQueries("CREATE TABLE trips(l long,s symbol index capacity 10, ts TIMESTAMP) timestamp(ts) partition by month;",
+                "insert into trips " +
+                        "  select x, case when x<=3 then 'ABC' when x>6 and x <= 9 then 'DEF' else 'GHI' end," +
+                        "  timestamp_sequence(to_timestamp('2022-01-03T00:00:00', 'yyyy-MM-ddTHH:mm:ss'), 100000000000) " +
+                        "  from long_sequence(10);");
+
+        assertQuery("l\ts\tts\n" +
+                        "9\tDEF\t2022-01-12T06:13:20.000000Z\n" +
+                        "8\tDEF\t2022-01-11T02:26:40.000000Z\n" +
+                        "7\tDEF\t2022-01-09T22:40:00.000000Z\n" +
+                        "3\tABC\t2022-01-05T07:33:20.000000Z\n" +
+                        "2\tABC\t2022-01-04T03:46:40.000000Z\n" +
+                        "1\tABC\t2022-01-03T00:00:00.000000Z\n",
+                "select l, s, ts from trips where s in (select 'DEF' union all select 'ABC' ) and length(s) = 3 order by ts desc",
+                null, null, true, false, false);
+    }
+
+    @Test
+    public void testOrderByDescWithFilterOnValuesRecordCursorFactory() throws Exception {
+        runQueries("CREATE TABLE trips(l long,s symbol index capacity 5, ts TIMESTAMP) timestamp(ts) partition by month;",
+                "insert into trips " +
+                        "  select x, 'A' || ( x%3 )," +
+                        "  timestamp_sequence(to_timestamp('2022-01-03T00:00:00', 'yyyy-MM-ddTHH:mm:ss'), 100000000000) " +
+                        "  from long_sequence(10);");
+        //A0, A1, A2, A0, A1, A2, A0, A1, A2, A0 
+        assertQuery("l\ts\tts\n" +
+                        "9\tA0\t2022-01-12T06:13:20.000000Z\n" +
+                        "8\tA2\t2022-01-11T02:26:40.000000Z\n" +
+                        "6\tA0\t2022-01-08T18:53:20.000000Z\n" +
+                        "5\tA2\t2022-01-07T15:06:40.000000Z\n" +
+                        "3\tA0\t2022-01-05T07:33:20.000000Z\n" +
+                        "2\tA2\t2022-01-04T03:46:40.000000Z\n",
+                "select l, s, ts from trips where s in ('A2', 'A0') and length(s) = 2 order by ts desc",
+                null, null, true, false, false);
+    }
+
+    @Test
+    public void testOrderByDescWithDataFrameRecordCursorFactory() throws Exception {
+        runQueries("CREATE TABLE trips(l long,s symbol index capacity 5, ts TIMESTAMP) timestamp(ts) partition by month;",
+                "insert into trips " +
+                        "  select x, 'A' || ( x%3 )," +
+                        "  timestamp_sequence(to_timestamp('2022-01-03T00:00:00', 'yyyy-MM-ddTHH:mm:ss'), 100000000000) " +
+                        "  from long_sequence(10);");
+        //A0, A1, A2, A0, A1, A2, A0, A1, A2, A0 
+        assertQuery("l\ts\tts\n" +
+                        "8\tA2\t2022-01-11T02:26:40.000000Z\n" +
+                        "5\tA2\t2022-01-07T15:06:40.000000Z\n" +
+                        "2\tA2\t2022-01-04T03:46:40.000000Z\n",
+                "select l, s, ts from trips where s = 'A2' and test_match() order by ts desc",
+                null, null, true, false, false);
+    }
+
+    @Test
+    public void testOrderByDescWithFilterOnExcludedValuesRecordCursorFactory() throws Exception {
+        runQueries("CREATE TABLE trips(l long,s symbol index capacity 5, ts TIMESTAMP) timestamp(ts) partition by month;",
+                "insert into trips " +
+                        "  select x, 'A' || ( x%3 )," +
+                        "  timestamp_sequence(to_timestamp('2022-01-03T00:00:00', 'yyyy-MM-ddTHH:mm:ss'), 100000000000) " +
+                        "  from long_sequence(10);");
+        assertQuery("l\ts\tts\n" +
+                        "9\tA0\t2022-01-12T06:13:20.000000Z\n" +
+                        "8\tA2\t2022-01-11T02:26:40.000000Z\n" +
+                        "6\tA0\t2022-01-08T18:53:20.000000Z\n" +
+                        "5\tA2\t2022-01-07T15:06:40.000000Z\n" +
+                        "3\tA0\t2022-01-05T07:33:20.000000Z\n" +
+                        "2\tA2\t2022-01-04T03:46:40.000000Z\n",
+                "select l, s, ts from trips where s != 'A1' and test_match() order by ts desc",
+                null, null, true, false, false);
+    }
+
+    @Test
+    public void testOrderByDescWithMultipleNotEqualsSymbolConditions() throws Exception {
+        runQueries("CREATE TABLE trips(l long,s symbol index capacity 5, ts TIMESTAMP) timestamp(ts) partition by month;",
+                "insert into trips " +
+                        "  select x, 'A' || ( x%3 )," +
+                        "  timestamp_sequence(to_timestamp('2022-01-03T00:00:00', 'yyyy-MM-ddTHH:mm:ss'), 100000000000) " +
+                        "  from long_sequence(10);");
+        assertQuery("l\ts\tts\n" +
+                        "8\tA2\t2022-01-11T02:26:40.000000Z\n" +
+                        "5\tA2\t2022-01-07T15:06:40.000000Z\n" +
+                        "2\tA2\t2022-01-04T03:46:40.000000Z\n",
+                "select l, s, ts from trips where s != 'A1' and s != 'A0' and test_match() order by ts desc",
+                null, null, true, false, false);
     }