This PR replaces sort by with full backward table scan for queries with filters (but without symbol predicates!) .
Example : 
````
select * from trips where total_amount > 100 order by pickup_datetime desc  limit 100
```
ran on demo dataset takes : 
in master - 17s (cold) and then 3.3 s (hot)
in this branch  - 42ms (cold) and then about 16 ms  (hot).

Note: there's a number of record cursor factories (e.g. DeferredSingleSymbolFilterDataFrameRecordCursorFactory) that don't support this optimization (yet) .