@ideoma 2022-02-02
#### [PR Coverage check]

:disappointed: **fail** : 6 / 23 (26.09%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/table/CompiledFilterRecordCursor.java|0|15|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/CompiledFilterRecordCursorFactory.java|0|2|00.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/FilteredRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|5|5|100.00%|