@ideoma 2022-02-03
#### [PR Coverage check]

:heart_eyes: **pass** : 24 / 24 (100.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/griffin/engine/table/FilteredRecordCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/CompiledFilterRecordCursorFactory.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/SequentialRowCursorFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/griffin/SqlCodeGenerator.java|5|5|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/table/CompiledFilterRecordCursor.java|15|15|100.00%|